+++ CHANGELOG.md
@@ -30,6 +30,7 @@
 * `Ctrl+y` binding to execute selected entry.
 * `Ctrl+j`/`Ctrl+k` binding to move to the next/previous item
   ([#120][120]).
+* Escape sequences in `Exec` arguments are now supported.
 
 [3]: https://codeberg.org/dnkl/fuzzel/issues/3
 [107]: https://codeberg.org/dnkl/fuzzel/issues/107